requirements.txt
hdfs
elasticsearch
beautifulsoup4
kafka-python

python-dotenv
schedule

pymongo
webdriver_manager
selenium